+++ .yipack/loader/babel-loader.config.js
@@ -1,30 +1,35 @@
-let myConfig = require("../webpack.config.my.js");
+let path = require('path');
+let myConfig = require('../webpack.config.my.js');
+
 module.exports = {
-    loader: "babel-loader",
+    loader: 'babel-loader',
     options: {
         cwd: myConfig.cliDir,
+        root: myConfig.cliDir,
+        rootMode: 'upward',
+        configFile: path.join(myConfig.cliDir, 'babel.config.js'),
         presets: [
             [
-                "@babel/preset-env",
+                '@babel/preset-env',
                 {
-                    useBuiltIns: "usage",
-                    corejs: "3",
-                },
-            ],
+                    useBuiltIns: 'usage',
+                    corejs: '3'
+                }
+            ]
         ],
         plugins: [
             [
-                "@babel/plugin-transform-runtime",
+                '@babel/plugin-transform-runtime',
                 {
                     absoluteRuntime: false,
                     corejs: 3,
                     helpers: true,
                     regenerator: true,
-                    useESModules: false,
-                },
+                    useESModules: false
+                }
             ],
-            "@babel/plugin-proposal-optional-chaining",
-            "@babel/plugin-proposal-nullish-coalescing-operator",
-        ],
-    },
+            '@babel/plugin-proposal-optional-chaining',
+            '@babel/plugin-proposal-nullish-coalescing-operator'
+        ]
+    }
 };