+++ babel.config.js
@@ -1,0 +1,2 @@
+let myConfig = require('./.yipack/webpack.config.my.js');
+module.exports = {};