+++ CHANGELOG.md
@@ -1,74 +1,305 @@
-# [2.2.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.1.1...v2.2.0) (2021-02-14)
+# [2.5.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.4.1...v2.5.0) (2021-03-01)
+
+
+### Features
+
+* 增加stylelint进行css验证 ([a7d307c](https://gitee.com/banshiweichen/yipack-cli/commits/a7d307c4ed4c67d5d58ca916ec752fe6ac21cb56))
+* 增加组件的属性排序规则 ([9317f59](https://gitee.com/banshiweichen/yipack-cli/commits/9317f5981b3f8645c42b91851a46d1d574d6e9e1))
+* 调整stylelint的缓存位置为.cache ([ef8a568](https://gitee.com/banshiweichen/yipack-cli/commits/ef8a568c9f282428dd1e6c6d98d1aa65e6483003))
+
+
+
+## [2.4.1](https://gitee.com/banshiweichen/yipack-cli/compare/v2.4.0...v2.4.1) (2021-03-01)
+
 
 ### Bug Fixes
 
--   修复 webpack profile 参数为 true 时的提示问题 ([a9b04ef](https://gitee.com/banshiweichen/yipack-cli/commits/a9b04efd971382a22e7a43d919bc42bed23cb39b))
--   增加代码风格和提交风格检测 ([ebdc1de](https://gitee.com/banshiweichen/yipack-cli/commits/ebdc1de0f290213d276432da61a20bb78591cf24))
+* 将eslint依赖放到项目依赖中 ([b840d9c](https://gitee.com/banshiweichen/yipack-cli/commits/b840d9c0e0e4818c354700599149570291a8c017))
+
+
+
+# [2.4.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.3.4...v2.4.0) (2021-02-28)
+
+
+### Bug Fixes
+
+* 修复创建全局组件时的组件目录名称 ([1b96709](https://gitee.com/banshiweichen/yipack-cli/commits/1b967095eae4454b2735cb8c8e2e4bc50b4e54e2))
+
+
+### Features
+
+* 成功将eslint相关依赖从项目移入到yipack-cli中 ([3a45aed](https://gitee.com/banshiweichen/yipack-cli/commits/3a45aed92e9671c3729c9d50ec49e8782e00790b))
+
 
 ### Performance Improvements
 
--   发布 2.2.0 ([159231c](https://gitee.com/banshiweichen/yipack-cli/commits/159231c9189aa59859cd1927d1c46e201aaf4289))
+* 更新相关依赖 ([ccb0bf0](https://gitee.com/banshiweichen/yipack-cli/commits/ccb0bf043e3b961692d067d678e7acb0fe54ce38))
+
+
+
+## [2.3.4](https://gitee.com/banshiweichen/yipack-cli/compare/v2.3.3...v2.3.4) (2021-02-27)
+
+
+### Bug Fixes
+
+* 修复接口自动生成的路径 ([0e85550](https://gitee.com/banshiweichen/yipack-cli/commits/0e85550d583768f1df3faa6c92b90c88764407ea))
+
+
+
+## [2.3.3](https://gitee.com/banshiweichen/yipack-cli/compare/v2.3.2...v2.3.3) (2021-02-27)
+
+
+### Bug Fixes
+
+* 修复页面的接口引入问题 ([68d32a0](https://gitee.com/banshiweichen/yipack-cli/commits/68d32a0baa1dcb31805558a1c83ba1665233a366))
+
+
+
+## [2.3.2](https://gitee.com/banshiweichen/yipack-cli/compare/v2.3.1...v2.3.2) (2021-02-25)
+
+
+### Features
+
+* 增加修复路由文件和内容 ([47b2c4d](https://gitee.com/banshiweichen/yipack-cli/commits/47b2c4d387732cc11ff859b07379c6b60707e7ce))
+
+
+
+## [2.3.1](https://gitee.com/banshiweichen/yipack-cli/compare/v2.3.0...v2.3.1) (2021-02-25)
+
+
+### Bug Fixes
+
+* 更新修复模板的地址 ([87b641f](https://gitee.com/banshiweichen/yipack-cli/commits/87b641fa3b4563c144aa881bd6a6c7f8a2726fe6))
+
+
+
+# [2.3.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.6...v2.3.0) (2021-02-25)
+
+
+### Features
+
+* 增加三级路由 ([b02dcb5](https://gitee.com/banshiweichen/yipack-cli/commits/b02dcb56da718d0f62b2a63a6192f5e6b1afbf53))
+
+
+### Performance Improvements
+
+* 小幅调整 ([0e56bc6](https://gitee.com/banshiweichen/yipack-cli/commits/0e56bc60a6c9c79a1c2b4188cb1961e9f38131b0))
+* 屏蔽.gitkeep文件的检测 ([6376d21](https://gitee.com/banshiweichen/yipack-cli/commits/6376d2169a8b0cb188634f024eef38a6f415dde0))
+
+
+
+## [2.2.6](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.5...v2.2.6) (2021-02-22)
+
+
+### Features
+
+* 完成fix命令 ([39ad1ac](https://gitee.com/banshiweichen/yipack-cli/commits/39ad1acc4247457aed4c5d7afe92d8634d9846e1))
+* 暂存fix --type=all命令 ([2b00eee](https://gitee.com/banshiweichen/yipack-cli/commits/2b00eee2c0555c04cc00f3aa8024d9d4bb44a832))
+
+
+
+## [2.2.5](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.4...v2.2.5) (2021-02-22)
+
+
+### Features
+
+* 增加全局组件和全局过滤器创建命令 ([e8392e5](https://gitee.com/banshiweichen/yipack-cli/commits/e8392e576411b7506828806a386aa3a01f836328))
+
+
+
+## [2.2.4](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.3...v2.2.4) (2021-02-22)
+
+
+### Features
+
+* 修改组件的创建方式为目录承载 ([1237f4b](https://gitee.com/banshiweichen/yipack-cli/commits/1237f4b9347368fbc5cad5a483ce72ed347332a9))
+
+
+
+## [2.2.3](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.2...v2.2.3) (2021-02-22)
+
+
+### Bug Fixes
+
+* 美化编译命令输出提示 ([716b4f2](https://gitee.com/banshiweichen/yipack-cli/commits/716b4f2ef0fd617d971005d84fb91833248681d1))
+
+
+### Features
+
+* 完成修复readme.md命令 ([14ff35d](https://gitee.com/banshiweichen/yipack-cli/commits/14ff35d08b8faa7cdc9504efeeb6a5c325a67f30))
+
+
+
+## [2.2.2](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.1...v2.2.2) (2021-02-21)
+
+
+### Bug Fixes
+
+* 修复自动创建页面模板时分号报错的问题 ([567d37b](https://gitee.com/banshiweichen/yipack-cli/commits/567d37bbfafacbea0c888888dec57a03a0691556))
+
+
+### Features
+
+* 创建命令完善 ([ee311c3](https://gitee.com/banshiweichen/yipack-cli/commits/ee311c3633b09ee94510aa255e29c9939447af82))
+* 默认提供on_init初始化方法 ([8087ac8](https://gitee.com/banshiweichen/yipack-cli/commits/8087ac8daf8ef675d241b1af030aa8416728fee6))
+
+
+
+## [2.2.1](https://gitee.com/banshiweichen/yipack-cli/compare/v2.2.0...v2.2.1) (2021-02-18)
+
+
+### Bug Fixes
+
+* 修改api导出的语法问题 ([9ad03cd](https://gitee.com/banshiweichen/yipack-cli/commits/9ad03cd3cfb2ba02e9f06fbe8f68a29d21c9e2ff))
+* 完善rename命令的缩写和全拼 ([cd51c72](https://gitee.com/banshiweichen/yipack-cli/commits/cd51c7200f28e9ee9cfe6a4f094635fb72285a5c))
+
+
+### Features
+
+* rename命令完成一半 ([d3f9309](https://gitee.com/banshiweichen/yipack-cli/commits/d3f93096b836ce3bf8917728177878288e450aed))
+
+
+
+# [2.2.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.1.1...v2.2.0) (2021-02-14)
+
+
+### Bug Fixes
+
+* 修复webpack profile参数为true时的提示问题 ([a9b04ef](https://gitee.com/banshiweichen/yipack-cli/commits/a9b04efd971382a22e7a43d919bc42bed23cb39b))
+* 增加代码风格和提交风格检测 ([ebdc1de](https://gitee.com/banshiweichen/yipack-cli/commits/ebdc1de0f290213d276432da61a20bb78591cf24))
+
+
+### Performance Improvements
+
+* 发布2.2.0 ([159231c](https://gitee.com/banshiweichen/yipack-cli/commits/159231c9189aa59859cd1927d1c46e201aaf4289))
+
+
 
 ## [2.1.1](https://gitee.com/banshiweichen/yipack-cli/compare/v2.1.0...v2.1.1) (2021-02-12)
 
+
+
 # [2.1.0](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.24...v2.1.0) (2021-01-29)
+
+
 
 ## [2.0.24](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.23...v2.0.24) (2021-01-28)
 
+
+
 ## [2.0.23](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.22...v2.0.23) (2021-01-13)
+
+
 
 ## [2.0.22](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.21...v2.0.22) (2021-01-11)
 
+
+
 ## [2.0.21](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.20...v2.0.21) (2021-01-05)
+
+
 
 ## [2.0.20](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.19...v2.0.20) (2021-01-05)
 
+
+
 ## [2.0.19](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.18...v2.0.19) (2021-01-05)
+
+
 
 ## [2.0.18](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.17...v2.0.18) (2021-01-04)
 
+
+
 ## [2.0.17](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.16...v2.0.17) (2021-01-04)
+
+
 
 ## [2.0.16](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.15...v2.0.16) (2021-01-04)
 
+
+
 ## [2.0.15](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.14...v2.0.15) (2020-12-25)
+
+
 
 ## [2.0.14](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.13...v2.0.14) (2020-12-16)
 
+
+
 ## [2.0.13](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.12...v2.0.13) (2020-12-16)
+
+
 
 ## [2.0.12](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.11...v2.0.12) (2020-12-16)
 
+
+
 ## [2.0.11](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.10...v2.0.11) (2020-12-15)
+
+
 
 ## [2.0.10](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.9...v2.0.10) (2020-12-15)
 
+
+
 ## [2.0.9](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.8...v2.0.9) (2020-12-14)
+
+
 
 ## [2.0.8](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.7...v2.0.8) (2020-12-10)
 
+
+
 ## [2.0.7](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.6...v2.0.7) (2020-12-06)
+
+
 
 ## [2.0.6](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.5...v2.0.6) (2020-12-06)
 
+
+
 ## [2.0.5](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.4...v2.0.5) (2020-11-30)
+
+
 
 ## [2.0.4](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.3...v2.0.4) (2020-11-30)
 
+
+
 ## [2.0.3](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.2...v2.0.3) (2020-11-30)
+
+
 
 ## [2.0.2](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.1...v2.0.2) (2020-11-26)
 
+
+
 ## [2.0.1](https://gitee.com/banshiweichen/yipack-cli/compare/v2.0.0...v2.0.1) (2020-11-26)
+
+
 
 # [2.0.0](https://gitee.com/banshiweichen/yipack-cli/compare/v1.0.7...v2.0.0) (2020-11-26)
 
+
+
 ## [1.0.7](https://gitee.com/banshiweichen/yipack-cli/compare/v1.0.6...v1.0.7) (2020-11-23)
+
+
 
 ## [1.0.6](https://gitee.com/banshiweichen/yipack-cli/compare/v1.0.5...v1.0.6) (2020-11-06)
 
+
+
 ## [1.0.5](https://gitee.com/banshiweichen/yipack-cli/compare/v1.0.4...v1.0.5) (2020-11-02)
+
+
 
 ## [1.0.4](https://gitee.com/banshiweichen/yipack-cli/compare/v1.0.3...v1.0.4) (2020-11-01)
 
+
+
 ## 1.0.3 (2020-10-18)
+
+
+